requirements.txt
SQLAlchemy~=2.0.30
httpx==0.23.3
fastapi==0.95.1
pytest~=8.2.1
asyncpg==0.27.0
greenlet==2.0.2
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
uvicorn==0.21.1
alembic~=1.13.1
psycopg2-binary==2.9.6
sqlalchemy_utils==0.40.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
starlette-exporter==0.17.1
google-api-python-client==2.133.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pyyaml==6.0.1

requests~=2.32.3
typing~=3.7.4.3
minio~=7.2.7
starlette~=0.26.1
pydantic~=1.10.15